fix(cdk/tree): fix broken examples in stackblitz
resolves #30159
